Flatten `Annotator` hierarchy.
