+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/annotation/FieldAnnotator.java
@@ -42,7 +42,7 @@ import java.nio.file.Path;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
-import static io.spine.tools.compiler.annotation.TypeDefinitionAnnotator.findNestedType;
+import static io.spine.tools.compiler.annotation.MessageAnnotator.findNestedType;
 import static io.spine.util.Exceptions.newIllegalStateException;
 
 /**
